Use a separate state to track the size of the bitmap brush tool from the brush tool
