+++ src/components/mode-tools/mode-tools.css
@@ -11,6 +11,12 @@
     margin-right: calc(2 * $grid-unit);
     width: 2rem;
     height: 2rem;
+
+    image-rendering: -moz-crisp-edges;
+    image-rendering: -o-crisp-edges;
+    image-rendering: -webkit-optimize-contrast;
+    -ms-interpolation-mode: nearest-neighbor;
+    image-rendering: pixelated;
 }
 
 .mod-dashed-border {
