+++ src/components/mode-tools/mode-tools.jsx
@@ -6,6 +6,7 @@ import React from 'react';
 
 import {changeBrushSize} from '../../reducers/brush-mode';
 import {changeBrushSize as changeEraserSize} from '../../reducers/eraser-mode';
+import {changeBitBrushSize} from '../../reducers/bit-brush-size';
 
 import LiveInputHOC from '../forms/live-input-hoc.jsx';
 import {defineMessages, injectIntl, intlShape} from 'react-intl';
@@ -13,12 +14,15 @@ import Input from '../forms/input.jsx';
 import InputGroup from '../input-group/input-group.jsx';
 import LabeledIconButton from '../labeled-icon-button/labeled-icon-button.jsx';
 import Modes from '../../lib/modes';
+import Formats from '../../lib/format';
+import {isBitmap} from '../../lib/format';
 import styles from './mode-tools.css';
 
 import copyIcon from './icons/copy.svg';
 import pasteIcon from './icons/paste.svg';
 
 import brushIcon from '../brush-mode/brush.svg';
+import bitBrushIcon from '../bit-brush-mode/brush.svg';
 import curvedPointIcon from './icons/curved-point.svg';
 import eraserIcon from '../eraser-mode/eraser.svg';
 import flipHorizontalIcon from './icons/flip-horizontal.svg';
@@ -75,6 +79,9 @@ const ModeToolsComponent = props => {
     switch (props.mode) {
     case Modes.BRUSH:
     case Modes.BIT_BRUSH:
+        const currentBrushIcon = isBitmap(props.format) ? bitBrushIcon : brushIcon;
+        const currentBrushValue = isBitmap(props.format) ? props.bitBrushSize : props.brushValue;
+        const changeFunction = isBitmap(props.format) ? props.onBitBrushSliderChange : props.onBrushSliderChange;
         return (
             <div className={classNames(props.className, styles.modeTools)}>
                 <div>
@@ -82,7 +89,7 @@ const ModeToolsComponent = props => {
                         alt={props.intl.formatMessage(messages.brushSize)}
                         className={styles.modeToolsIcon}
                         draggable={false}
-                        src={brushIcon}
+                        src={currentBrushIcon}
                     />
                 </div>
                 <LiveInput
@@ -91,8 +98,8 @@ const ModeToolsComponent = props => {
                     max={MAX_STROKE_WIDTH}
                     min="1"
                     type="number"
-                    value={props.brushValue}
-                    onSubmit={props.onBrushSliderChange}
+                    value={currentBrushValue}
+                    onSubmit={changeFunction}
                 />
             </div>
         );
@@ -175,15 +182,18 @@ const ModeToolsComponent = props => {
 };
 
 ModeToolsComponent.propTypes = {
+    bitBrushSize: PropTypes.number,
     brushValue: PropTypes.number,
     className: PropTypes.string,
     clipboardItems: PropTypes.arrayOf(PropTypes.array),
     eraserValue: PropTypes.number,
+    format: PropTypes.oneOf(Object.keys(Formats)).isRequired,
     hasSelectedUncurvedPoints: PropTypes.bool,
     hasSelectedUnpointedPoints: PropTypes.bool,
     intl: intlShape.isRequired,
     mode: PropTypes.string.isRequired,
-    onBrushSliderChange: PropTypes.func,
+    onBitBrushSliderChange: PropTypes.func.isRequired,
+    onBrushSliderChange: PropTypes.func.isRequired,
     onCopyToClipboard: PropTypes.func.isRequired,
     onCurvePoints: PropTypes.func.isRequired,
     onEraserSliderChange: PropTypes.func,
@@ -196,6 +206,8 @@ ModeToolsComponent.propTypes = {
 
 const mapStateToProps = state => ({
     mode: state.scratchPaint.mode,
+    format: state.scratchPaint.format,
+    bitBrushSize: state.scratchPaint.bitBrushSize,
     brushValue: state.scratchPaint.brushMode.brushSize,
     clipboardItems: state.scratchPaint.clipboard.items,
     eraserValue: state.scratchPaint.eraserMode.brushSize,
@@ -204,6 +216,9 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch => ({
     onBrushSliderChange: brushSize => {
         dispatch(changeBrushSize(brushSize));
+    },
+    onBitBrushSliderChange: bitBrushSize => {
+        dispatch(changeBitBrushSize(bitBrushSize));
     },
     onEraserSliderChange: eraserSize => {
         dispatch(changeEraserSize(eraserSize));
