+++ src/helper/bit-tools/brush-tool.js
@@ -1,5 +1,6 @@
 import paper from '@scratch/paper';
 import {getRaster} from '../layer';
+import {line, fillEllipse} from '../bitmap';
 
 /**
  * Tool for drawing with the bitmap brush.
@@ -29,84 +30,6 @@ class BrushTool extends paper.Tool {
         // For performance, make sure this is an integer
         this.size = ~~size;
     }
-    line (point1, point2, callback){
-        // Bresenham line algorithm
-        // Fast Math.floor
-        let x1 = ~~point1.x;
-        const x2 = ~~point2.x;
-        let y1 = ~~point1.y;
-        const y2 = ~~point2.y;
-        
-        const dx = Math.abs(x2 - x1);
-        const dy = Math.abs(y2 - y1);
-        const sx = (x1 < x2) ? 1 : -1;
-        const sy = (y1 < y2) ? 1 : -1;
-        let err = dx - dy;
-        
-        callback(x1, y1);
-        while (x1 !== x2 || y1 !== y2) {
-            const e2 = err * 2;
-            if (e2 > -dy) {
-                err -= dy; x1 += sx;
-            }
-            if (e2 < dx) {
-                err += dx; y1 += sy;
-            }
-            callback(x1, y1);
-        }
-    }
-    fillEllipse (centerX, centerY, radiusX, radiusY, context) {
-        // Bresenham ellipse algorithm
-        centerX = ~~centerX;
-        centerY = ~~centerY;
-        radiusX = ~~radiusX;
-        radiusY = ~~radiusY;
-        const twoRadXSquared = 2 * radiusX * radiusX;
-        const twoRadYSquared = 2 * radiusY * radiusY;
-        let x = radiusX;
-        let y = 0;
-        let dx = radiusY * radiusY * (1 - (radiusX << 1));
-        let dy = radiusX * radiusX;
-        let error = 0;
-        let stoppingX = twoRadYSquared * radiusX;
-        let stoppingY = 0;
-     
-        while (stoppingX >= stoppingY) {
-            context.fillRect(centerX - x, centerY - y, x << 1, y << 1);
-            y++;
-            stoppingY += twoRadXSquared;
-            error += dy;
-            dy += twoRadXSquared;
-            if ((error << 1) + dx > 0) {
-                x--;
-                stoppingX -= twoRadYSquared;
-                error += dx;
-                dx += twoRadYSquared;
-            }
-        }
-
-        x = 0;
-        y = radiusY;
-        dx = radiusY * radiusY;
-        dy = radiusX * radiusX * (1 - (radiusY << 1));
-        error = 0;
-        stoppingX = 0;
-        stoppingY = twoRadXSquared * radiusY;
-        while (stoppingX <= stoppingY) {
-            context.fillRect(centerX - x, centerY - y, x * 2, y * 2);
-            x++;
-            stoppingX += twoRadYSquared;
-            error += dx;
-            dx += twoRadYSquared;
-            if ((error << 1) + dy > 0) {
-                y--;
-                stoppingY -= twoRadXSquared;
-                error += dy;
-                dy += twoRadXSquared;
-            }
-
-        }
-    }
     // Draw a brush mark at the given point
     draw (x, y) {
         getRaster().drawImage(this.tmpCanvas, new paper.Point(x - ~~(this.size / 2), y - ~~(this.size / 2)));
@@ -125,7 +48,7 @@ class BrushTool extends paper.Tool {
         if (this.size <= 4) {
             context.fillRect(0, 0, this.size, this.size);
         } else {
-            this.fillEllipse(this.size / 2, this.size / 2, this.size / 2, this.size / 2, context);
+            fillEllipse(this.size / 2, this.size / 2, this.size / 2, this.size / 2, context);
         }
 
         this.draw(event.point, event.point);
@@ -138,13 +61,14 @@ class BrushTool extends paper.Tool {
             this.boundingBoxTool.onMouseDrag(event);
             return;
         }
-        this.line(this.lastPoint, event.point, this.draw.bind(this));
+        line(this.lastPoint, event.point, this.draw.bind(this));
         this.lastPoint = event.point;
     }
     handleMouseUp (event) {
         if (event.event.button > 0 || !this.active) return; // only first mouse button
         
-        this.line(this.lastPoint, event.point, this.draw.bind(this));
+        line(this.lastPoint, event.point, this.draw.bind(this));
+        this.onUpdateSvg();
 
         this.tmpCanvas = null;
         this.lastPoint = null;
