+++ src/helper/bitmap.js
@@ -1,5 +1,85 @@
 import paper from '@scratch/paper';
 
+const line = function (point1, point2, callback) {
+    // Bresenham line algorithm
+    // Fast Math.floor
+    let x1 = ~~point1.x;
+    const x2 = ~~point2.x;
+    let y1 = ~~point1.y;
+    const y2 = ~~point2.y;
+    
+    const dx = Math.abs(x2 - x1);
+    const dy = Math.abs(y2 - y1);
+    const sx = (x1 < x2) ? 1 : -1;
+    const sy = (y1 < y2) ? 1 : -1;
+    let err = dx - dy;
+    
+    callback(x1, y1);
+    while (x1 !== x2 || y1 !== y2) {
+        const e2 = err * 2;
+        if (e2 > -dy) {
+            err -= dy; x1 += sx;
+        }
+        if (e2 < dx) {
+            err += dx; y1 += sy;
+        }
+        callback(x1, y1);
+    }
+};
+
+const fillEllipse = function (centerX, centerY, radiusX, radiusY, context) {
+    // Bresenham ellipse algorithm
+    centerX = ~~centerX;
+    centerY = ~~centerY;
+    radiusX = ~~radiusX;
+    radiusY = ~~radiusY;
+    const twoRadXSquared = 2 * radiusX * radiusX;
+    const twoRadYSquared = 2 * radiusY * radiusY;
+    let x = radiusX;
+    let y = 0;
+    let dx = radiusY * radiusY * (1 - (radiusX << 1));
+    let dy = radiusX * radiusX;
+    let error = 0;
+    let stoppingX = twoRadYSquared * radiusX;
+    let stoppingY = 0;
+ 
+    while (stoppingX >= stoppingY) {
+        context.fillRect(centerX - x, centerY - y, x << 1, y << 1);
+        y++;
+        stoppingY += twoRadXSquared;
+        error += dy;
+        dy += twoRadXSquared;
+        if ((error << 1) + dx > 0) {
+            x--;
+            stoppingX -= twoRadYSquared;
+            error += dx;
+            dx += twoRadYSquared;
+        }
+    }
+
+    x = 0;
+    y = radiusY;
+    dx = radiusY * radiusY;
+    dy = radiusX * radiusX * (1 - (radiusY << 1));
+    error = 0;
+    stoppingX = 0;
+    stoppingY = twoRadXSquared * radiusY;
+    while (stoppingX <= stoppingY) {
+        context.fillRect(centerX - x, centerY - y, x * 2, y * 2);
+        x++;
+        stoppingX += twoRadYSquared;
+        error += dx;
+        dx += twoRadYSquared;
+        if ((error << 1) + dy > 0) {
+            y--;
+            stoppingY -= twoRadXSquared;
+            error += dy;
+            dy += twoRadXSquared;
+        }
+
+    }
+};
+
 const rowBlank_ = function (imageData, width, y) {
     for (let x = 0; x < width; ++x) {
         if (imageData.data[(y * width << 2) + (x << 2) + 3] !== 0) return false;
@@ -33,5 +113,7 @@ const trim = function (raster) {
 };
 
 export {
+    fillEllipse,
+    line,
     trim
 };
