+++ src/reducers/scratch-paint-reducer.js
@@ -1,5 +1,6 @@
 import {combineReducers} from 'redux';
 import modeReducer from './modes';
+import bitBrushSizeReducer from './bit-brush-size';
 import brushModeReducer from './brush-mode';
 import eraserModeReducer from './eraser-mode';
 import colorReducer from './color';
@@ -14,6 +15,7 @@ import undoReducer from './undo';
 
 export default combineReducers({
     mode: modeReducer,
+    bitBrushSize: bitBrushSizeReducer,
     brushMode: brushModeReducer,
     color: colorReducer,
     clipboard: clipboardReducer,
